Completed the comments and on Hover, they are functional, but not saved as JSON.
I overcomplicated the save data, it worked how I wanted it to work in the beginning, so I changed it back. The on hover is functional, but could be better. Will have to mess with CSS to improve it. Comments are currently saved into the Item class object Rating. The comments will appear on hover of the rating button. Clicking the rating button will delete the comment at that spot allowing recomment of that item. Saving comment to JSON should be handled by the submit button. Item class should have most of information required.